added list practice
